Resume campaign immediately when (re-)adding users (#197)
* Resume campaign when (re-)adding users

Fix a bug where the campaign fails to resume immediately when adding a
user who had previously received earlier steps in the campaign.

See this issue comment for additional info:

https://github.com/honeybadger-io/heya/issues/179#issuecomment-1185978092

Background: when a user is added to a campaign [^1], we create a new
`CampaignMembership`. The two attributes that control scheduling are
`step_gid` (the current step we're waiting for) and `last_sent_at` (the
time a step was last successfully processed). When a user matches a
segment and an email is sent, we create a `CampaignReceipt` which stores
a permanent record of that step being sent to that user. **When a user
exits the campaign, their `CampaignMembership` is destroyed, but the
`CampaignReceipt`(s) stay.** This is what enables Heya to skip steps
[^2] if the user is re-added later.

I think the bug is that Heya isn't checking for existing
`CampaignReceipt`(s) early enough, and is waiting to process those steps
in the meantime. Basically they will still be skipped, but it's waiting
to skip them until the scheduler tries to process them (which is after
their `wait` time).

As you can see here [^3], the user always starts a campaign on step 1.
What it needs to do is find the first step for which a `CampaignReceipt`
does not exist, and start there. If there is no `CampaignReceipt` for
the user+campaign, then it should start on step 1 like it currently
does.

That should fix *most* of this issue, but there is one other gotcha:
`last_sent_at` defaults to the time the user was last added to the
campaign [^4]. That means that if the user exits a campaign and then you
add a step and re-add them a month from now, it will wait whatever the
`wait` time is for that step. It would be better if `last_sent_at` would
default to whatever it was when they previously exited the campaign, but
of course, the previous `CampaignMembership` was destroyed. What we
*could* do is again, look for the last prior `CampaignReceipt` for the
user+campaign, and if one exists, do something like this:
`self.last_sent_at = last_campaign_receipt.created_at`. That way it
should be able to pick up right where it left off. 🤞

[^1]: https://github.com/honeybadger-io/heya/blob/7fd40b2638469ef21646880fc9da1bfa980dab54/lib/heya/campaigns/base.rb#L45
[^2]: https://github.com/honeybadger-io/heya/blob/7fd40b2638469ef21646880fc9da1bfa980dab54/lib/heya/campaigns/scheduler.rb#L38
[^3]: https://github.com/honeybadger-io/heya/blob/7fd40b2638469ef21646880fc9da1bfa980dab54/lib/heya/campaigns/base.rb#L44
[^4]: https://github.com/honeybadger-io/heya/blob/7fd40b2638469ef21646880fc9da1bfa980dab54/app/models/heya/campaign_membership.rb#L8

* Jump forward to next undelivered step in scheduler

See https://github.com/honeybadger-io/heya/issues/179#issuecomment-1185987257

* Update CHANGELOG.md

* Compare integer dates instead?

+++ lib/heya/campaigns/base.rb
@@ -19,7 +19,7 @@ module Heya
       end
 
       delegate :name, :__segments, to: :class
-      alias id name
+      alias_method :id, :name
 
       # Returns String GlobalID.
       def gid
@@ -35,16 +35,18 @@ module Heya
           membership.delete_all
         end
 
-        if restart
-          CampaignReceipt
-            .where(user: user, step_gid: steps.map(&:gid))
-            .delete_all
-        end
+        receipts = CampaignReceipt
+          .where(user: user, step_gid: steps.map(&:gid))
 
-        if (step = steps.first)
+        receipts.delete_all if restart
+
+        last_sent_at = receipts.maximum(:created_at)
+        receipt_gids = receipts.map(&:step_gid)
+        if (step = steps.find { |s| receipt_gids.exclude?(s.gid) })
           membership.create! do |m|
             m.concurrent = concurrent
             m.step_gid = step.gid
+            m.last_sent_at = last_sent_at
           end
 
           if send_now && step.wait == 0
@@ -111,7 +113,7 @@ module Heya
         end
 
         def segment(arg = nil, &block)
-          if block_given?
+          if block
             self.__segments = ([block] | __segments).freeze
           elsif arg
             self.__segments = ([arg] | __segments).freeze
@@ -119,7 +121,7 @@ module Heya
         end
 
         def step(name, **opts, &block)
-          if block_given?
+          if block
             opts[:block] ||= block
             opts[:action] ||= Actions::Block
           end
